refactor: Update imports for location modals to match component file names
