+++ src/components/modules/location/CreateEquipmentModal.tsx
