+++ src/components/modules/location/UpdateEquipmentModal.tsx
