+++ src/pages/locations/index.tsx
@@ -11,12 +11,12 @@ import { LayoutApp } from '@/components/layout/LayoutApp'
 import { Input } from '@/components/elements/Input'
 import { LocationDropDown } from '@/components/modules/location/LocationDropDown'
 import { CreateLocationPopoverPanelOption } from '@/components/modules/location/CreateLocationPopoverPanelOption'
-import { CreateAreaModal } from '@/components/modules/location/CreateArea'
-import { CreateConstructionModal } from '@/components/modules/location/CreateConstruction'
-import { CreateEquipmentModal } from '@/components/modules/location/CreateEquipment'
-import { UpdateAreaModal } from '@/components/modules/location/UpdateArea'
-import { UpdateConstructionModal } from '@/components/modules/location/UpdateConstruction'
-import { UpdateEquipmentModal } from '@/components/modules/location/UpdateEquipment'
+import { CreateAreaModal } from '@/components/modules/location/CreateAreaModal'
+import { CreateConstructionModal } from '@/components/modules/location/CreateConstructionModal'
+import { CreateEquipmentModal } from '@/components/modules/location/CreateEquipmentModal'
+import { UpdateAreaModal } from '@/components/modules/location/UpdateAreaModal'
+import { UpdateConstructionModal } from '@/components/modules/location/UpdateConstructionModal'
+import { UpdateEquipmentModal } from '@/components/modules/location/UpdateEquipmentModal'
 import { DeleteLocationModal } from '@/components/modules/location/DeleteLocationModal'
 
 const Location: NextPage = () => {
